+++ gradio/utils.py
@@ -2,7 +2,6 @@
 
 from __future__ import annotations
 
-import ast
 import asyncio
 import copy
 import functools
@@ -79,6 +78,7 @@ from gradio.themes import ThemeClass as Theme
 if TYPE_CHECKING:  # Only import for type checking (is False at runtime).
     from gradio.blocks import BlockContext, Blocks
     from gradio.components import Component
+    from gradio.components.button import Button
     from gradio.routes import App, Request
     from gradio.state_holder import SessionState
 
@@ -155,6 +155,17 @@ class BaseReloader(ABC):
     def running_app(self) -> App:
         pass
 
+    def get_attribute(self, attr: str, demo) -> Any:
+        if (
+            hasattr(demo, f"_deprecated_{attr}")
+            and getattr(demo, f"_deprecated_{attr}") is not None
+        ):
+            return getattr(demo, f"_deprecated_{attr}")
+        elif hasattr(demo, attr) and getattr(demo, attr) is not None:
+            return getattr(demo, attr)
+        else:
+            return getattr(self.running_app.blocks, attr)
+
     def swap_blocks(self, demo: Blocks):
         assert self.running_app.blocks  # noqa: S101
         # Copy over the blocks to get new components and events but
@@ -165,11 +176,13 @@ class BaseReloader(ABC):
         demo.is_running = True
         demo.allowed_paths = self.running_app.blocks.allowed_paths
         demo.blocked_paths = self.running_app.blocks.blocked_paths
-        demo.theme = self.running_app.blocks.theme
-        demo.head_paths = self.running_app.blocks.head_paths
-        demo.css = self.running_app.blocks.css
-        demo.head = self.running_app.blocks.head
-        demo.css_paths = self.running_app.blocks.css_paths
+
+        demo.theme = self.get_attribute("theme", demo)
+        demo.css = self.get_attribute("css", demo)
+        demo.css_paths = self.get_attribute("css_paths", demo)
+        demo.js = self.get_attribute("js", demo)
+        demo.head = self.get_attribute("head", demo)
+        demo.head_paths = self.get_attribute("head_paths", demo)
         demo._set_html_css_theme_variables()
         self.running_app.state_holder.set_blocks(demo)
         for session in self.running_app.state_holder.session_data.values():
@@ -294,51 +307,6 @@ class SourceFileReloader(ServerReloader):
         self.alert_change("reload")
 
 
-def _remove_if_name_main_codeblock(file_path: str, encoding: str = "utf-8"):
-    """Parse the file, remove the gr.no_reload code blocks, and write the file back to disk.
-
-    Parameters:
-        file_path (str): The path to the file to remove the no_reload code blocks from.
-    """
-
-    with open(file_path, encoding=encoding) as file:
-        code = file.read()
-
-    tree = ast.parse(code)
-
-    def _is_if_name_main(expr: ast.AST) -> bool:
-        """Find the if __name__ == '__main__': block."""
-        return (
-            isinstance(expr, ast.If)
-            and isinstance(expr.test, ast.Compare)
-            and isinstance(expr.test.left, ast.Name)
-            and expr.test.left.id == "__name__"
-            and len(expr.test.ops) == 1
-            and isinstance(expr.test.ops[0], ast.Eq)
-            and isinstance(expr.test.comparators[0], ast.Constant)
-            and expr.test.comparators[0].s == "__main__"
-        )
-
-    # Find the positions of the code blocks to load
-    for node in ast.walk(tree):
-        if _is_if_name_main(node):
-            assert isinstance(node, ast.If)  # noqa: S101
-            node.body = [ast.Pass(lineno=node.lineno, col_offset=node.col_offset)]
-
-    # convert tree to string
-    code_removed = compile(tree, filename=file_path, mode="exec")
-    return code_removed
-
-
-def _find_module(source_file: Path) -> ModuleType | None:
-    for s, v in sys.modules.items():
-        if s not in {"__main__", "__mp_main__"} and getattr(v, "__file__", None) == str(
-            source_file
-        ):
-            return v
-    return None
-
-
 def watchfn_spaces(reloader: SpacesReloader):
     try:
         spaces_version = importlib.metadata.version("spaces")
@@ -422,8 +390,8 @@ def watchfn(reloader: SourceFileReloader) -> None:
     # Need to import the module in this thread so that the
     # module is available in the namespace of this thread
     module = reloader.watch_module
-    no_reload_source_code = _remove_if_name_main_codeblock(
-        str(reloader.demo_file), encoding=reloader.encoding
+    no_reload_source_code = Path(str(reloader.demo_file)).read_text(
+        encoding=reloader.encoding
     )
     # Reset the context to id 0 so that the loaded module is the same as the original
     # See https://github.com/gradio-app/gradio/issues/10253
@@ -456,8 +424,8 @@ def watchfn(reloader: SourceFileReloader) -> None:
 
                 NO_RELOAD.set(False)
                 # Remove the gr.no_reload code blocks and exec in the new module's dict
-                no_reload_source_code = _remove_if_name_main_codeblock(
-                    str(reloader.demo_file), encoding=reloader.encoding
+                no_reload_source_code = Path(str(reloader.demo_file)).read_text(
+                    encoding=reloader.encoding
                 )
                 exec(no_reload_source_code, module.__dict__)
 
@@ -1965,3 +1933,16 @@ async def safe_aclose_iterator(iterator, timeout=60.0, retry_interval=0.05):
                     raise
     else:
         iterator.aclose()
+
+
+def set_default_buttons(
+    buttons: Sequence[str | Button] | None = None,
+    default_buttons: list[str] | None = None,
+) -> Sequence[str | Button]:
+    from gradio.components.button import Button
+
+    if buttons is None:
+        return default_buttons or []
+    else:
+        [btn.unrender() for btn in buttons if isinstance(btn, Button)]
+        return buttons