+++ js/audio/Index.svelte
@@ -147,6 +147,10 @@
 			waveform_options={gradio.props.waveform_options}
 			editable={gradio.props.editable}
 			{minimal}
+			on_custom_button_click={(id) => {
+				gradio.dispatch("custom_button_click", { id });
+			}}
+			bind:playback_position={gradio.props.playback_position}
 			on:share={(e) => gradio.dispatch("share", e.detail)}
 			on:error={(e) => gradio.dispatch("error", e.detail)}
 			on:play={() => gradio.dispatch("play")}
@@ -182,6 +186,9 @@
 			{label}
 			show_label={gradio.shared.show_label}
 			buttons={gradio.props.buttons ?? []}
+			on_custom_button_click={(id) => {
+				gradio.dispatch("custom_button_click", { id });
+			}}
 			value={gradio.props.value}
 			subtitles={gradio.props.subtitles}
 			on:change={({ detail }) => (gradio.props.value = detail)}
@@ -201,6 +208,7 @@
 			{handle_reset_value}
 			editable={gradio.props.editable}
 			bind:dragging
+			bind:playback_position={gradio.props.playback_position}
 			on:edit={() => gradio.dispatch("edit")}
 			on:play={() => gradio.dispatch("play")}
 			on:pause={() => gradio.dispatch("pause")}