+++ js/audio/shared/types.ts
@@ -1,4 +1,5 @@
 import { FileData } from "@gradio/client";
+import type { CustomButton } from "@gradio/utils";
 
 export type WaveformOptions = {
 	waveform_color?: string;
@@ -25,7 +26,7 @@ export interface AudioProps {
 	value: FileData | null;
 	type: "numpy" | "filepath";
 	autoplay: boolean;
-	buttons: ("play" | "download")[];
+	buttons: ("play" | "download" | "share" | CustomButton)[];
 	recording: boolean;
 	loop: boolean;
 	subtitles: FileData | SubtitleData[] | null;
@@ -36,6 +37,7 @@ export interface AudioProps {
 	stream_every: number;
 	input_ready: boolean;
 	minimal?: boolean;
+	playback_position: number;
 }
 
 export interface AudioEvents {
@@ -55,4 +57,5 @@ export interface AudioEvents {
 	clear_status: any;
 	close_stream: any;
 	edit: any;
+	custom_button_click: { id: number };
 }