+++ js/checkboxgroup/types.ts
@@ -1,10 +1,11 @@
-import type { SelectData } from "@gradio/utils";
+import type { SelectData, CustomButton } from "@gradio/utils";
 
 export interface CheckboxGroupProps {
 	value: (string | number)[];
 	choices: [string, string | number][];
 	info: string;
 	show_select_all: boolean;
+	buttons: (string | CustomButton)[] | null;
 }
 
 export interface CheckboxGroupEvents {
@@ -12,4 +13,5 @@ export interface CheckboxGroupEvents {
 	input: (string | number)[];
 	select: SelectData;
 	clear_status: void;
+	custom_button_click: { id: number };
 }