+++ js/fileexplorer/types.ts
@@ -1,4 +1,5 @@
 import type { LoadingStatus } from "@gradio/statustracker";
+import type { SelectData, CustomButton } from "@gradio/utils";
 import type { FileNode } from "./shared/types";
 
 export interface FileExplorerProps {
@@ -10,12 +11,17 @@ export interface FileExplorerProps {
 	root_dir: string;
 	glob: string;
 	ignore_glob: string;
+	_selectable: boolean;
 	server: {
 		ls: (path: string[]) => Promise<FileNode[]>;
 	};
+	buttons: (string | CustomButton)[] | null;
 }
 
 export interface FileExplorerEvents {
 	change: never;
+	input: never;
+	select: SelectData;
 	clear_status: LoadingStatus;
+	custom_button_click: { id: number };
 }