+++ js/video/Index.svelte
@@ -100,10 +100,11 @@
 			show_label={gradio.shared.show_label}
 			autoplay={gradio.props.autoplay}
 			loop={gradio.props.loop}
-			show_share_button={(gradio.props.buttons || []).includes("share")}
-			show_download_button={(gradio.props.buttons || ["download"]).includes(
-				"download"
-			)}
+			buttons={gradio.props.buttons ?? ["download", "share"]}
+			on_custom_button_click={(id) => {
+				gradio.dispatch("custom_button_click", { id });
+			}}
+			bind:playback_position={gradio.props.playback_position}
 			on:play={() => gradio.dispatch("play")}
 			on:pause={() => gradio.dispatch("pause")}
 			on:stop={() => gradio.dispatch("stop")}
@@ -149,7 +150,10 @@
 			bind:uploading
 			label={gradio.shared.label}
 			show_label={gradio.shared.show_label}
-			show_download_button={(gradio.props.buttons || []).includes("download")}
+			buttons={gradio.props.buttons ?? ["download", "share"]}
+			on_custom_button_click={(id) => {
+				gradio.dispatch("custom_button_click", { id });
+			}}
 			sources={gradio.props.sources}
 			{active_source}
 			webcam_options={gradio.props.webcam_options}
@@ -158,6 +162,7 @@
 			root={gradio.shared.root}
 			loop={gradio.props.loop}
 			{handle_reset_value}
+			bind:playback_position={gradio.props.playback_position}
 			on:clear={() => {
 				gradio.props.value = null;
 				gradio.dispatch("clear");